+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🐞 Fixed
 - Fix incorrect RawJSON number handling, the `.integer` case is no longer supported and is replaced by `.number` [#1375](https://github.com/GetStream/stream-chat-swift/pull/1375)
+- Fix message list and thread index out of range issue on `tableView(_:cellForRowAt:)` [1373](https://github.com/GetStream/stream-chat-swift/pull/1373)
 
 ### 🔄 Changed